Make raw item bank Float64

+++ src/dummydata.jl
@@ -11,7 +11,7 @@ function clumpy_4pl_item_bank(rng, num_clumps, num_questions)
     )
     params_clumps = mapslices(Product, clump_dist_mat; dims=[2])[:, 1]
     # TODO: Resample the clumps to create a correlated distribution
-    params = Array{Float32, 2}(undef, num_questions, 4)
+    params = Array{Float64, 2}(undef, num_questions, 4)
     for (question_idx, clump) in enumerate(sample(rng, params_clumps, num_questions; replace=true))
         (difficulty, discrimination, guess, slip) = rand(rng, clump)
         params[question_idx, :] = [difficulty, abs(discrimination), pclamp(guess), pclamp(slip)]
